requirements.txt
langchain>=0.1.0
chromadb>=0.4.22
langchain-google-genai>=0.0.5
google-genai
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.2
